+++ udata_front/theme/gouvfr/datagouv-components/demo/App.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
-import { Pagination, ResourceAccordion, Well } from "../src";
+import { Pagination, ResourceAccordion, QualityComponent, QualityComponentInline, QualityItem, QualityScore, ReadMore, Toggletip, Well } from "../src";
 import { ref } from "vue";
 import type { Resource } from "../src";
+import type { Quality } from "../src/types/datasets";
 import editIcon from "../../templates/svg/illustrations/edit.svg";
 
 const resource = ref<Resource>({
@@ -78,6 +79,21 @@ test.. test... test..... test?..... test!....
     organization: null,
     url: "https://static.data.gouv.fr/resources/indice-de-reparabilite-organisation-ribimex/20231115-104022/data.csv"
   });
+
+  const quality = ref<Quality>({
+    all_resources_available: false,
+    dataset_description_quality: false,
+    has_open_format: true,
+    has_resources: false,
+    license: false,
+    resources_documentation: true,
+    score: 0.9,
+    spatial: false,
+    temporal_coverage: false,
+    update_frequency: false,
+    update_fulfilled_in_time: false,
+  });
+  const id = "12";
 </script>
 
 <template>
@@ -86,6 +102,32 @@ test.. test... test..... test?..... test!....
   </h1>
   <ResourceAccordion dataset-id="someId" :resource="resource" :expanded-on-mount="false" />
   <ResourceAccordion dataset-id="someId" :resource="resourceWithoutSchema" :expanded-on-mount="false" />
+
+  <QualityScore :score="0.5" class="w-100"></QualityScore>
+  <QualityItem :passed="true" message-passed="Oui" message-failed="Non"></QualityItem>
+  <Toggletip>
+    Toggle
+    <template #toggletip>
+      <h5 class="fr-text--sm fr-my-0">
+      {{$t("Metadata quality:")}}
+      </h5>
+      <QualityItem :passed="true" message-passed="Oui" message-failed="Non"></QualityItem>
+    </template>
+  </Toggletip>
+  <QualityComponent 
+    :quality="quality"
+  ></QualityComponent>
+  <QualityComponentInline
+    :quality="quality"
+  ></QualityComponentInline>
+  <ReadMore>
+    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin aliquet, justo vitae facilisis lobortis, turpis velit suscipit urna, vel rutrum arcu lacus vitae felis. Proin ac orci et tellus auctor sodales. Cras eleifend a lacus tempor tristique. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae; Integer mi urna, tempor vel feugiat eget, aliquet quis eros. Sed nisi tortor, fringilla vitae orci vel, faucibus imperdiet mauris. Maecenas nunc eros, dapibus vitae orci ut, semper ultricies sem.
+
+    Pellentesque nulla ex, condimentum quis urna quis, ultrices vehicula risus. Duis facilisis nisl massa, et consequat nisi tristique ultricies. Praesent aliquam bibendum urna, non fermentum augue tempor ac. Sed tincidunt nibh sapien, a posuere risus mollis sit amet. Cras lorem massa, fringilla pulvinar dapibus ac, pharetra eu enim. Nam arcu ante, consequat vitae ultricies non, posuere vitae quam. Etiam imperdiet, ligula non venenatis suscipit, dolor leo mattis quam, dignissim sodales tortor felis vitae turpis. Ut sed congue tellus, fringilla scelerisque leo. Fusce viverra varius convallis. Aliquam faucibus odio nunc, malesuada gravida sem gravida sed. Vivamus sollicitudin dolor euismod libero aliquam, at cursus sapien auctor.
+
+    Pellentesque nulla ex, condimentum quis urna quis, ultrices vehicula risus. Duis facilisis nisl massa, et consequat nisi tristique ultricies. Praesent aliquam bibendum urna, non fermentum augue tempor ac. Sed tincidunt nibh sapien, a posuere risus mollis sit amet. Cras lorem massa, fringilla pulvinar dapibus ac, pharetra eu enim. Nam arcu ante, consequat vitae ultricies non, posuere vitae quam. Etiam imperdiet, ligula non venenatis suscipit, dolor leo mattis quam, dignissim sodales tortor felis vitae turpis. Ut sed congue tellus, fringilla scelerisque leo. Fusce viverra varius convallis. Aliquam faucibus odio nunc, malesuada gravida sem gravida sed. Vivamus sollicitudin dolor euismod libero aliquam, at cursus sapien auctor.</p>
+  </ReadMore>
+
   <Well color="blue-cumulus" weight="regular" class="fr-my-2w">
     <div class="fr-grid-row">
       <div class="fr-col-auto fr-mr-3v">