+++ udata_front/theme/gouvfr/datagouv-components/src/components/QualityComponent/index.ts
@@ -1,0 +1,2 @@
+import QualityComponent from "./QualityComponent.vue";
+export { QualityComponent };