+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Move Well to datagouv-components [#382](https://github.com/etalab/udata-front/pull/382)
 - Add markdown editor [#351](https://github.com/etalab/udata-front/pull/351)
 - Upgrade vue dependency [#386](https://github.com/etalab/udata-front/pull/386)
+- Componentize quality component [#385](https://github.com/etalab/udata-front/pull/385)
 - Add codes and optgroups in Multiselect to display Insee codes and Licence groups [#347] (https://github.com/etalab/udata-front/pull/347)
 - Fix datastructure display in ResourceAccordion [#389](https://github.com/etalab/udata-front/pull/389)
 - Order Organization's reuses by publishing date [#390](https://github.com/etalab/udata-front/pull/390)